refactor: 迁移共享展示层基础设施
将 BaseViewModel 与 UiEvent 从共享数据模型目录迁移至 app/presentation，明确其展示层职责；不改变运行时行为。
