+++ lib/features/app_update/models/event.dart
@@ -1,4 +1,4 @@
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 
 class AppUpdateMigrationLinkCopiedEvent extends UiEvent {
   const AppUpdateMigrationLinkCopiedEvent();
