+++ lib/features/app_update/view_models/app_update_migration_view_model.dart
@@ -3,8 +3,8 @@ import 'package:flutter/services.dart';
 import 'package:onetj/app/di/dependencies.dart';
 import 'package:onetj/app/logging/logger.dart';
 import 'package:onetj/features/app_update/models/event.dart';
-import 'package:onetj/models/base_model.dart';
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/base_view_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 import 'package:onetj/services/external_launcher_service.dart';
 
 class AppUpdateMigrationViewModel extends BaseViewModel<UiEvent> {
