+++ lib/features/app_update/views/app_update_migration_dialog.dart
@@ -7,7 +7,7 @@ import 'package:onetj/app/constant/app_version_constant.dart';
 import 'package:onetj/features/app_update/models/event.dart';
 import 'package:onetj/features/app_update/view_models/app_update_migration_view_model.dart';
 import 'package:onetj/models/app_update_info.dart';
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 
 class AppUpdateMigrationDialog extends StatefulWidget {
   const AppUpdateMigrationDialog({
