+++ lib/features/grades/views/grades_view.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:onetj/l10n/app_localizations.dart';
 
 import 'package:onetj/features/grades/view_models/grades_view_model.dart';
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 
 class GradesView extends StatefulWidget {
   const GradesView({super.key});
