+++ lib/features/launcher/view_models/launcher_view_model.dart
@@ -3,8 +3,8 @@ import 'package:onetj/app/di/dependencies.dart';
 import 'package:onetj/app/logging/logger.dart';
 import 'package:onetj/app/logging/logging_bootstrap.dart';
 import 'package:onetj/app/theme/theme_change_notifier.dart';
-import 'package:onetj/models/base_model.dart';
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/base_view_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 import 'package:onetj/models/launch_wallpaper_ref.dart';
 import 'package:onetj/models/settings_data.dart';
 import 'package:onetj/repo/settings_repository.dart';
