+++ lib/features/launcher/views/launcher_view.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
 import 'package:onetj/features/launcher/view_models/launcher_view_model.dart';
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 import 'package:onetj/models/settings_defaults.dart';
 
 class LauncherView extends StatefulWidget {
