+++ lib/features/login/views/login_view.dart
@@ -5,7 +5,7 @@ import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:onetj/l10n/app_localizations.dart';
 import 'package:go_router/go_router.dart';
 
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 import 'package:onetj/app/exception/app_exception.dart';
 
 import '../view_models/login_view_model.dart';
