+++ lib/features/settings/views/log_viewer_view.dart
@@ -8,7 +8,7 @@ import 'package:path/path.dart' as path;
 import 'package:onetj/app/logging/log_file_info.dart';
 import 'package:onetj/features/settings/models/event.dart';
 import 'package:onetj/features/settings/view_models/log_viewer_view_model.dart';
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 
 class LogViewerView extends StatefulWidget {
   const LogViewerView({super.key});
