+++ lib/features/settings/views/settings_view.dart
@@ -17,7 +17,7 @@ import 'package:onetj/features/settings/views/widgets/theme_color_card.dart';
 import 'package:onetj/features/settings/views/widgets/home_layout_card.dart';
 import 'package:onetj/features/settings/views/widgets/upcoming_courses_card.dart';
 import 'package:onetj/models/dashboard_upcoming_mode.dart';
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 import 'package:onetj/models/launch_wallpaper_ref.dart';
 import 'package:onetj/models/time_period_range.dart';
 import 'package:onetj/models/time_slot.dart';
