+++ lib/features/timetable/models/event.dart
@@ -1,4 +1,4 @@
-import 'package:onetj/models/event_model.dart';
+import 'package:onetj/app/presentation/ui_event.dart';
 
 class SyncWheelEvent extends UiEvent {
   const SyncWheelEvent();
